# src/AasRegistryDiscovery.App/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "DotAAS Part 2 | HTTP/REST | Asset Administration Shell Registry Service Specification",
    "description" : "The Full Profile of the Asset Administration Shell Registry Service Specification as part of the [Specification of the Asset Administration Shell: Part 2](http://industrialdigitaltwin.org/en/content-hub).  \nPublisher: Industrial Digital Twin Association (IDTA) 2023",
    "contact" : {
      "name" : "Industrial Digital Twin Association (IDTA)",
      "email" : "info@idtwin.org"
    },
    "license" : {
      "name" : "CC BY 4.0",
      "url" : "https://creativecommons.org/licenses/by/4.0/"
    },
    "version" : "V3.0.3_SSP-001",
    "x-profile-identifier" : "https://admin-shell.io/aas/API/3/0/AssetAdministrationShellRegistryServiceSpecification/SSP-001"
  },
  "servers" : [ {
    "url" : "{protocol}://{host_name}:{port}/api/{version_prefix}",
    "variables" : {
      "protocol" : {
        "description" : "Allows access through http and https (recommended)",
        "default" : "https",
        "enum" : [ "http", "https" ]
      },
      "host_name" : {
        "description" : "Hostname of server hosting the api",
        "default" : "admin-shell.io"
      },
      "port" : {
        "description" : "80 is default for http, 443 for https",
        "default" : "443",
        "enum" : [ "80", "443" ]
      },
      "version_prefix" : {
        "default" : "v3.0",
        "enum" : [ "v3.0" ]
      }
    }
  } ],
  "paths" : {
    "/shell-descriptors" : {
      "get" : {
        "tags" : [ "Asset Administration Shell Registry API" ],
        "summary" : "Returns all Asset Administration Shell Descriptors",
        "operationId" : "GetAllAssetAdministrationShellDescriptors",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "description" : "The maximum number of elements in the response array",
          "required" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "name" : "cursor",
          "in" : "query",
          "description" : "A server-generated identifier retrieved from pagingMetadata that specifies from which position the result listing should continue",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "assetKind",
          "in" : "query",
          "description" : "The Asset's kind (Instance or Type)",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/AssetKind"
          }
        }, {
          "name" : "assetType",
          "in" : "query",
          "description" : "The Asset's type (UTF8-BASE64-URL-encoded)",
          "required" : false,
          "schema" : {
            "maxLength" : 2000,
            "minLength" : 1,
            "pattern" : "^([\\x09\\x0a\\x0d\\x20-\\ud7ff\\ue000-\\ufffd]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Requested Asset Administration Shell Descriptors",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GetAssetAdministrationShellDescriptorsResult"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request, e.g. the request parameters of the format of the request body is wrong.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "default" : {
            "description" : "Default error handling for unmentioned status codes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/GetAllAssetAdministrationShellDescriptors/3/0", "https://admin-shell.io/aas/API/GetAllAssetAdministrationShellDescriptorsByAssetKind/3/0", "https://admin-shell.io/aas/API/GetAllAssetAdministrationShellDescriptorsByAssetType/3/0" ]
      },
      "post" : {
        "tags" : [ "Asset Administration Shell Registry API" ],
        "summary" : "Creates a new Asset Administration Shell Descriptor, i.e. registers an AAS",
        "operationId" : "PostAssetAdministrationShellDescriptor",
        "requestBody" : {
          "description" : "Asset Administration Shell Descriptor object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AssetAdministrationShellDescriptor"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Asset Administration Shell Descriptor created successfully",
            "headers" : {
              "Location" : {
                "description" : "URL of the newly created resource",
                "style" : "simple",
                "explode" : false,
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AssetAdministrationShellDescriptor"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request, e.g. the request parameters of the format of the request body is wrong.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "403" : {
            "description" : "Forbidden",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "409" : {
            "description" : "Conflict, a resource which shall be created exists already. Might be thrown if a Submodel or SubmodelElement with the same ShortId is contained in a POST request.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          },
          "default" : {
            "description" : "Default error handling for unmentioned status codes",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Result"
                }
              }
            }
          }
        },
        "x-semanticIds" : [ "https://admin-shell.io/aas/API/PostAssetAdministrationShellDescriptor/3/0" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "AssetKind" : {
        "type" : "string",
        "enum" : [ "Instance", "NotApplicable", "Type" ]
      },
      "GetAssetAdministrationShellDescriptorsResult" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/PagedResult"
        }, {
          "type" : "object",
          "properties" : {
            "result" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/AssetAdministrationShellDescriptor"
              }
            }
          }
        } ]
      },
      "PagedResult" : {
        "type" : "object",
        "properties" : {
          "paging_metadata" : {
            "$ref" : "#/components/schemas/PagedResult_paging_metadata"
          }
        }
      },
      "AssetAdministrationShellDescriptor" : {
        "required" : [ "id" ],
        "type" : "object",
        "example" : "{ \"id\": \"https://example.org/aas/motor\", \"endpoints\": [ { \"protocolInformation\": { \"href\": \"https://localhost:1234/api/v3.0/aas\", \"endpointProtocol\": \"HTTP\", \"endpointProtocolVersion\": [\"1.1\"] }, \"interface\": \"AAS-3.0\" }, { \"protocolInformation\": { \"href\": \"opc.tcp://localhost:4840\" }, \"interface\": \"AAS-3.0\" }, { \"protocolInformation\": { \"href\": \"https://localhost:5678\", \"endpointProtocol\": \"HTTP\", \"endpointProtocolVersion\": [\"1.1\"], \"subprotocol\": \"OPC UA Basic SOAP\", \"subprotocolBody\": \"ns=2;s=MyAAS\", \"subprotocolBodyEncoding\": \"application/soap+xml\" }, \"interface\": \"AAS-3.0\" } ] }",
        "allOf" : [ {
          "$ref" : "#/components/schemas/Descriptor"
        }, {
          "properties" : {
            "administration" : {
              "$ref" : "#/components/schemas/AdministrativeInformation"
            },
            "assetKind" : {
              "$ref" : "#/components/schemas/AssetKind"
            },
            "assetType" : {
              "maxLength" : 2000,
              "minLength" : 1,
              "pattern" : "^([\\x09\\x0a\\x0d\\x20-\\ud7ff\\ue000-\\ufffd]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "endpoints" : {
              "minItems" : 1,
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Endpoint"
              }
            },
            "globalAssetId" : {
              "maxLength" : 2000,
              "minLength" : 1,
              "pattern" : "^([\\x09\\x0a\\x0d\\x20-\\ud7ff\\ue000-\\ufffd]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "idShort" : {
              "maxLength" : 128,
              "type" : "string"
            },
            "id" : {
              "maxLength" : 2000,
              "minLength" : 1,
              "pattern" : "^([\\x09\\x0a\\x0d\\x20-\\ud7ff\\ue000-\\ufffd]|\\ud800[\\udc00-\\udfff]|[\\ud801-\\udbfe][\\udc00-\\udfff]|\\udbff[\\udc00-\\udfff])*$",
              "type" : "string"
            },
            "specificAssetIds" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/SpecificAssetId"
              }
            },
            "submodelDescriptors" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/SubmodelDescriptor"
              }
            }
          }
        } ]
      },
...
      "Endpoint" : {
        "required" : [ "interface", "protocolInformation" ],
        "type" : "object",
        "properties" : {
          "interface" : {
            "maxLength" : 128,
            "type" : "string"
          },
          "protocolInformation" : {
            "$ref" : "#/components/schemas/ProtocolInformation"
          }
        }
      },
...
      "Result" : {
        "type" : "object",
        "properties" : {
          "messages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Message"
            }
          }
        }
      },
...
    },
...
  }
}